server/requirements.txt
flask
flask-cors
python-dotenv
cohere
pytube
google-cloud-speech
google-cloud-storage